fix: audit cleanup — fix bugs, remove dead code, deduplicate utilities
- Fix stale closure in handleSelection causing unnecessary SwipeCard re-renders
- Fix stale onDismiss callback in MatchToast auto-dismiss timer
- Fix showSwipeHint missing from useEffect deps in SwipeCard
- Fix timer leak in BubblePillsBackground and WelcomeSplash spawn chains
- Remove 5 unused animated styles/callbacks from useCardAnimation hook
- Extract duplicate GENDER_EMOJI constant into constants/names.ts
- Extract duplicate getRelativeTime into lib/format.ts
- Merge near-identical LikedNamesHeader/RejectedNamesHeader into shared NamesHeader
- Fix wrong gender keys (boy/girl/unisex -> male/female/neutral) in MatchCard
