+++ components/dashboard/rejected-name-card.tsx
@@ -1,6 +1,8 @@
 import { View, Text, Pressable, StyleSheet, Alert } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { Fonts } from '@/constants/theme';
+import { GENDER_EMOJI } from '@/constants/names';
+import { getRelativeTime } from '@/lib/format';
 import { getOriginFlag } from '@/constants/origins';
 import { useTheme } from '@/contexts/theme-context';
 import { Doc } from '@/convex/_generated/dataModel';
@@ -14,42 +16,6 @@ interface RejectedNameCardProps {
   selectMode?: boolean;
   selected?: boolean;
   onToggleSelect?: () => void;
-}
-
-const GENDER_EMOJI: Record<string, string> = {
-  male: '👦',
-  female: '👧',
-  neutral: '👶',
-};
-
-
-function getRelativeTime(timestamp: number): string {
-  const now = Date.now();
-  const diff = now - timestamp;
-
-  const seconds = Math.floor(diff / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-  const weeks = Math.floor(days / 7);
-  const months = Math.floor(days / 30);
-
-  if (months > 0) {
-    return `${months} month${months > 1 ? 's' : ''} ago`;
-  }
-  if (weeks > 0) {
-    return `${weeks} week${weeks > 1 ? 's' : ''} ago`;
-  }
-  if (days > 0) {
-    return `${days} day${days > 1 ? 's' : ''} ago`;
-  }
-  if (hours > 0) {
-    return `${hours} hour${hours > 1 ? 's' : ''} ago`;
-  }
-  if (minutes > 0) {
-    return `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
-  }
-  return 'Just now';
 }
 
 export function RejectedNameCard({
