+++ constants/names.ts
@@ -1,0 +1,5 @@
+export const GENDER_EMOJI: Record<string, string> = {
+  male: '\u{1F466}',
+  female: '\u{1F467}',
+  neutral: '\u{1F476}',
+};
